+++ tasmota/language/de_DE.h
@@ -1271,6 +1271,8 @@
 #define D_SENSOR_PIPSOLAR_RX             "Pipsolar RX"
 
 // xsns_115_wooliis.ino
+#define D_IMPORT                          "Import"
+#define D_EXPORT                          "Export"
 #define D_CHARGING                        "Aufladen"
 #define D_CAPACITY                        "Kapazität"